warnings lua added ...
* emptied control.lua a little more, into the new wanrings.lua, and moved some other functions to appropriate modules.
* more can be emptied out
